[Refactor] ProductRender > ProductRenderPerRow로 이름 변경 및 리팩토링
