+++ src/components/common/ProductRenderPerRow.module.css
@@ -5,7 +5,6 @@
   justify-content: space-between;
   flex-wrap: nowrap;
   gap: 24px ;
-
 }
 
 .ProductBox {
@@ -64,7 +63,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  gap: 1px;
+  gap: 3.25px;
 }
 
 .ProductLike > span {
